Remove (normally) uneeded GitHub properties
The githubPush trigger is not needed for multibranch pipelines. It is
still needed on single pipeline jobs, but in that case we require the
project itself to use this. This is due to avoid a conflict merging
external pipelineTriggers with the ones in buildConfig, which e.g.
made cron trigger not work.

The GitHub project is not needed either as it will cause a duplicate
GitHub link and not be related to the actual branch the job belongs to.