insideMaven: Remove mount for .m2
The mount has actually never worked, as the HOME inside the container
for a maven image is always set to / when using pipelines (due to not
currently being able to override HOME when running a container).

This mount did however cause issues on classic slaves, where the
/home/jenkins/.m2 folder is actually used. If a modern job is run on
the classic slaves first, Docker will create the .m2 folder as root,
and then a classic job running later cannot access this directory
due to ownership issues.

The setup is now changed so that the injected settings.xml has a
localRepository property instead, pointing to
/home/jenkins/.m2/repository.